requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
asgiref==3.9.2
attrs==25.1.0
Beaker==1.12.1
beautifulsoup4==4.13.5
blinker==1.9.0
blivet==3.12.1
blivet-gui==2.6.0
boto3==1.40.33
botocore==1.40.33
Brlapi==0.8.6
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cockpit @ file:///builddir/build/BUILD/cockpit-347-build/cockpit-347/tmp/wheel/cockpit-347-py3-none-any.whl#sha256=8816ee14baaafcd61279d6180ad0c57fd497ac6ad5d73db9bc0de1fa66d9e21d
crypt_r==3.13.1
cryptography==44.0.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
dbus_next==0.2.3
distlib==0.3.9
distro==1.9.0
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnf==4.23.0
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.18.0
Flask==3.1.2
fros==1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.0
idna==3.10
iniconfig==1.1.1
iso639==0.1.4
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langtable==0.0.69
libcomps==0.1.22
libdnf==0.74.0
louis==3.33.0
lxml==5.3.2
Mako==1.2.3
MarkupSafe==3.0.2
mccabe==0.7.0
nftables==0.1
olefile==0.47
openai==1.84.0
packaging==24.2
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
platformdirs==4.3.8
pluggy==1.5.0
ply==3.11
productmd==1.45
ptyprocess==0.7.0
pwquality==1.4.5
PyAudio==0.2.13
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.11.5
pydantic_core==2.33.2
pyenchant==3.2.2
PyGObject==3.50.0
pyinotify==0.9.6
PyJWT==2.10.1
pykickstart==3.62
pyOpenSSL==25.0.0
pyparted==3.13.0
PySocks==1.7.1
pytest==8.3.4
python-augeas==1.2.0
python-dateutil==2.8.2
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.16.8-build/kernel-6.16.8/linux-6.16.8-200.fc42.x86_64/tools/net/ynl
RapidFuzz==3.11.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.0.0
requests-ftp==0.3.1
rpds-py==0.25.0
rpm==4.20.1
s3transfer==0.14.0
selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
setools==4.5.1
setuptools==74.1.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sos==4.10.0
soupsieve==2.7
sqlparse==0.5.3
systemd-python==235
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.31.2
Werkzeug==3.1.3
xkbregistry==0.3
